- Full fine-grained authentication support added - Fixed all db path compares in httpd match() - server can now serve default html files from the www directory
git-svn-id: http://caffeine-hx.googlecode.com/svn/trunk@327 5be69f5d-4a45-0410-b242-ad22a5f24ccc

+++ memedb/src/java/memedb/httpd/DeleteDocument.java
@@ -31,8 +31,14 @@ import memedb.document.Document;
  */
 public class DeleteDocument extends BaseRequestHandler {
 
-	public void handleInner(Credentials credentials, HttpServletRequest request, HttpServletResponse response, String db, String id, String rev) throws IOException, BackendException{
+	public void handleInner(Credentials credentials, HttpServletRequest request, HttpServletResponse response, String db, String id, String rev) throws IOException, BackendException
+	{
+		if(!credentials.canDeleteDocuments(db)) {
+			this.sendNotAuth(response);
+			return;
+		}
 		try{
+			//memeDB.getBackend().doesDatabaseExist(db)
 			memeDB.getBackend().deleteDocument(db,id);
 		}
 		catch(BackendException e) {}
@@ -40,7 +46,10 @@ public class DeleteDocument extends BaseRequestHandler {
 	}
 
 	public boolean match(Credentials credentials, HttpServletRequest request, String db, String id) {
-		return (db!=null && id!=null && request.getMethod().equals("DELETE") && memeDB.getBackend().doesDatabaseExist(db) && credentials.isAuthorizedWrite(db));
+		return (
+				db!=null && 
+				id!=null && 
+				request.getMethod().equals("DELETE"));
 	}
 
 }
